# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Team Management API
  version: 1.0.0
  description: |
    The contract both backends (`backend-python/`, `backend-node/`) implement and
    the frontend consumes. BetterDev's hidden checks call the running server over
    HTTP, so **paths, status codes and response shapes here are the spec** — the
    internals are yours.

    Conventions
    - JSON everywhere, camelCase fields, integer ids, ISO-8601 UTC timestamps.
    - Auth: `Authorization: Bearer <token>` from `POST /auth/login`. Every route
      except `/health`, `/auth/login` and `/__test__/*` requires it (401 otherwise).
    - Errors always use the `Error` envelope below.
    - List endpoints are paginated: `limit` (default 20, max 100) and `offset`
      (default 0), response `{items, total, limit, offset}`. The activity feed is
      the exception — it uses cursor pagination.
    - Roles: every user has a global `role` (`admin` | `member`). Teams have
      `lead` | `member`; projects have `admin` | `member`.
    - Visibility: a project (and its tasks, activity and numbers) is visible to its
      members and to global admins. For anyone else it does not exist — `404`, never
      `403`. `403` is for callers who can see the resource but may not change it.
    - Every response carries `X-Request-ID` (the incoming one when the client sent
      it, otherwise a generated one). Built in be-observability.

    Status markers: `x-status: shipped` = implemented in the starter (possibly
...
servers:
...
security:
  - bearer: []
paths:
...
  /__test__/reset:
    post:
      x-status: shipped
      security: []
      summary: Reset the database to a seed (only when BD_TEST_MODE=1; 404 otherwise)
      description: |
        Drops all data and loads `contracts/seed.json` (`size=small`) or the
        generated large dataset (`size=large`). Used by the hidden checks — do not
        remove it. When you add new tables, extend the seeder so they load too.
      parameters:
        - in: query
          name: size
          schema: { enum: [small, large], default: small }
      responses:
        '204': { description: Reset done }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /users:
    get:
      x-status: shipped
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: search
          description: |
            Case-insensitive substring match on email, first name, last name,
            "first last" or display name. Filtered and paginated in the database —
            `total` counts every match, `items` holds at most `limit` of them.
          schema: { type: string }
        - in: query
          name: role
          schema: { $ref: '#/components/schemas/UserRole' }
      responses:
        '200':
          description: Page of users ordered by id
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
    post:
      x-status: shipped
      summary: Create a user (global admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      x-status: shipped
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      x-status: shipped
      summary: Update a user (self, or global admin). Only admins may change `role`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      x-status: shipped
      summary: Delete a user (global admin only)
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /teams:
    get:
      x-status: shipped
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Page of teams (with members) ordered by id
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Team' }
    post:
      x-status: shipped
      summary: Create a team (global admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /teams/{teamId}:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    get:
      x-status: shipped
      responses:
        '200':
          description: Team
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      x-status: shipped
      summary: Update a team (global admin or team lead)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      x-status: shipped
      summary: Delete a team (global admin only)
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /teams/{teamId}/members:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    post:
      x-status: shipped
      summary: Add a member (global admin or team lead)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: integer }
                role: { $ref: '#/components/schemas/TeamRole' }
      responses:
        '201':
          description: Added — returns the updated team
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: User is already a member
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /teams/{teamId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/TeamId'
      - $ref: '#/components/parameters/UserId'
    delete:
      x-status: shipped
      summary: Remove a member (global admin or team lead)
      responses:
        '204': { description: Removed }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      in: query
      name: offset
      schema: { type: integer, minimum: 0, default: 0 }
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: integer }
    TeamId:
      in: path
      name: teamId
      required: true
      schema: { type: integer }
...
  responses:
...
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflicts with existing state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Invalid input — `details` lists the fields
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              enum: [unauthorized, forbidden, not_found, conflict, validation_error, rate_limited, internal_error]
            message: { type: string }
            details:
              description: For validation_error — one entry per invalid field
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
    Page:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    UserRole: { enum: [admin, member] }
    TeamRole: { enum: [lead, member], default: member }
...
    User:
      type: object
      required: [id, email, firstName, lastName, displayName, role, createdAt, updatedAt]
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      description: Never includes the password or its hash.
    UserCreate:
      type: object
      required: [email, firstName, lastName, password]
      properties:
        email: { type: string, format: email }
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        displayName: { type: string, maxLength: 160, description: 'Defaults to "firstName lastName"' }
        password: { type: string, minLength: 8 }
        role: { $ref: '#/components/schemas/UserRole' }
    UserUpdate:
      type: object
      minProperties: 1
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        displayName: { type: string, minLength: 1, maxLength: 160 }
        role: { $ref: '#/components/schemas/UserRole' }
    TeamMember:
      type: object
      required: [userId, displayName, role]
      properties:
        userId: { type: integer }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/TeamRole' }
    Team:
      type: object
      required: [id, name, description, members, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        members:
          type: array
          description: Ordered by userId
          items: { $ref: '#/components/schemas/TeamMember' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TeamCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120, description: Unique (case-insensitive) }
        description: { type: string, maxLength: 2000, default: '' }
    TeamUpdate:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 2000 }
...